Seperate out color, material, and presets into color/material/camera files respectively
